Remove the relationship between `PromoCampaign`s and creatives.
Previously there was an identical creative for each `PromoCampaign`
since creatives are actually derived from links.  This simplifies
things by only having 1 creative per link and gets us closer to the
model we want to use in the future.  It also reduces the amount of
copying to adzerk (and the potential for things to become out of
sync).